requirements.txt
opencv-python
streamlit>=1.27.0
transformers>=4.39.0
sentencepiece>=0.1.99
torch>=2.0.0
SpeechRecognition>=3.8.1
pyaudio==0.2.11
pytesseract>=0.3.10
Pillow>=9.5.0
langdetect>=1.0.9
sacremoses  
